requirements.txt
argh==0.26.2
asn1crypto==0.24.0
astroid==1.6.1
attrs==17.4.0
backports.functools-lru-cache==1.5
backports.tempfile==1.0
backports.weakref==1.0.post1
bcrypt==3.1.4
blessings==1.6.1
bpython==0.17.1
certifi==2018.1.18
cffi==1.11.4
chardet==3.0.4
citizenshell==1.1.3
colorama==0.3.9
configparser==3.5.0
cryptography==2.1.4
curtsies==0.3.0
docopt==0.6.2
enum34==1.1.6
funcsigs==1.0.2
futures==3.2.0
greenlet==0.4.13
idna==2.6
ipaddress==1.0.19
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
paramiko==2.4.0
pathtools==0.1.2
pkginfo==1.4.1
pluggy==0.6.0
py==1.5.2
pyasn1==0.4.2
pycparser==2.18
Pygments==2.2.0
pylint==1.8.2
PyNaCl==1.2.1
pypandoc==1.4
pyserial==3.4
pytest==3.4.0
pytest-sugar==0.9.1
pytest-watch==4.1.0
PyYAML==3.12
requests==2.18.4
requests-toolbelt==0.8.0
singledispatch==3.4.0.3
six==1.11.0
termcolor==1.1.0
tqdm==4.19.5
twine==1.9.1
typing==3.6.4
uritools==2.1.0
urllib3==1.22
watchdog==0.8.3
wcwidth==0.1.7
wrapt==1.10.11
